% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Add two numbers}
\usage{
foo(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
A number.
}
\description{
Adds \code{a} and \code{b} together and returns the sum.
}
\examples{
foo(1, 2)
}
